componentize CSS and cleanup inline tachyons
