+++ client/src/styles/App.scss
@@ -1,5 +1,10 @@
 @import "node_modules/tachyons-sass/tachyons.scss";
+@import "node_modules/@fortawesome/fontawesome-free/scss/fontawesome.scss";
+@import "node_modules/@fortawesome/fontawesome-free/scss/regular.scss";
+@import "node_modules/@fortawesome/fontawesome-free/scss/solid.scss";
 @import "./colors";
 @import "./variables";
 @import "./extend-tachyons";
-@import "./market";
+@import "./icons";
+@import "./pages/market";
+@import "./pages/dashboard";
