+++ client/src/styles/_extend-tachyons.scss
@@ -1,4 +1,9 @@
 // Grid extensions
+.grid-1-1 {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+}
+
 .grid-3 {
     display: grid;
     grid-template-columns: 1fr 1fr 1fr;
