--- client/src/styles/_market.scss
@@ -1,65 +1,0 @@
-.stalk-market {
-    .card {
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-
-        .trend {
-            .peak-price {
-                background-color: var(--primary-1);
-
-                .star {
-                    clip-path: polygon(
-                        50% 0%,
-                        61% 35%,
-                        98% 35%,
-                        68% 57%,
-                        79% 91%,
-                        50% 70%,
-                        21% 91%,
-                        32% 57%,
-                        2% 35%,
-                        39% 35%
-                    );
-                }
-
-                .star-large {
-                    @extend .star;
-                    width: 1.5rem;
-                    height: 1.5rem;
-                    background-color: var(--secondary-1);
-                }
-
-                .star-placeholder {
-                    @extend .star-large;
-                    width: 0;
-                    background-color: transparent;
-                }
-
-                .star-small {
-                    @extend .star;
-                    width: 1.25rem;
-                    height: 1.25rem;
-                    background-color: var(--secondary-1);
-                }
-            }
-        }
-
-        .icon-positive,
-        .icon-negative,
-        .icon-neutral {
-            @extend .mr2;
-        }
-
-        .icon-positive {
-            @extend .dark-green;
-        }
-
-        .icon-negative {
-            color: var(--tertiary-1);
-        }
-
-        .icon-neutral {
-            @extend .mid-gray;
-        }
-    }
-}
